feat(audit-p2): reference FKs + one-of CHECKs, invoice snapshot, integer qty, dead-field removal
Phase 2 batch (migration 0075 applied to prod: additive, pre-verified 0 CHECK
violations, backfilled 32 POs + 21 GRNs):
- Reference linkage: added purchase_orders.pr_id (PR->PO forward FK) and
  goods_receipt_notes.delivery_challan_id (GRN->DC forward FK) + backfill + set
  going-forward in createPurchaseOrderFromPr / insertGrnForOspReceipt. One-of
  CHECK backstops on PO vendor, PO-line item, GRN-line item.
- Invoices: removed the dead LEFT JOIN items (item_id always null) — invoice
  lines now read the frozen text snapshot only (correct legal-doc behavior; no
  live-master leak). Per product decision (freeze snapshot).
- Delivery Challan qty tightened to integer (product decision; 0 fractional rows
  on prod).
- Removed dead outsourceLeadDays across plan schema/service/forms (never reached
  jc_ops). Drizzle column def kept for a follow-up drop migration.

Workspace typecheck + lint green. Disjoint-file multi-agent, agents rejected 2
mis-scoped removals (edit.tsx/edit-plan-modal.tsx refs) rather than break typecheck.

+++ packages/shared/src/schemas/delivery-challan.ts
@@ -134,7 +134,7 @@ export const createDeliveryChallanLineInputSchema = z.object({
   itemId: z.string().uuid().nullable().optional(),
   itemCodeText: z.string().min(1),
   itemNameText: z.string().nullable().optional(),
-  qty: z.number().positive(),
+  qty: z.number().int().positive(),
   uom: uomSchema,
   materialText: z.string().nullable().optional(),
   dcRemarks: z.string().nullable().optional(),
@@ -206,8 +206,8 @@ export type DeliveryChallanReceipt = z.infer<typeof deliveryChallanReceiptSchema
 export const createDeliveryChallanReceiptLineInputSchema = z
   .object({
     deliveryChallanLineId: z.string().uuid(),
-    receivedQty: z.number().nonnegative(),
-    rejectedQty: z.number().nonnegative().default(0),
+    receivedQty: z.number().int().nonnegative(),
+    rejectedQty: z.number().int().nonnegative().default(0),
     rejectReason: z.string().nullable().optional(),
     remarks: z.string().nullable().optional(),
   })
